// src/lib/api/generated/model/friendRequest.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * Etechs Social Network
 * API contract cho social-network-ui (main API + message service).
 * OpenAPI spec version: 1.0
 */
import type { Author } from './author';
import type { FriendRequestStatus } from './friendRequestStatus';

export interface FriendRequest {
  id: string;
  requester: Author;
  addressee: Author;
  status: FriendRequestStatus;
  created_at: string;
}